refactor: Room 코드 리팩토링
